services/support/requirements.txt
annotated-types==0.7.0

anyio==4.9.0

bcrypt==4.3.0

beautifulsoup4==4.13.4

blinker==1.9.0

certifi==2025.7.9

cffi==1.17.1

charset-normalizer==3.4.2

click==8.2.1

colorama==0.4.6

cryptography==45.0.5

dnspython==2.7.0

fastapi==0.116.1

Flask==3.1.1

flask-cors==6.0.1

h11==0.16.0

idna==3.10

itsdangerous==2.2.0

Jinja2==3.1.6

MarkupSafe==3.0.2

paramiko==3.5.1

pycparser==2.22

pydantic==2.11.7

pydantic_core==2.33.2

PyNaCl==1.5.0

python-dateutil==2.9.0.post0

python-multipart==0.0.20

python-whois==0.9.5

requests==2.32.4

six==1.17.0

sniffio==1.3.1

soupsieve==2.7

starlette==0.47.2

typing-inspection==0.4.1

typing_extensions==4.14.1

urllib3==2.5.0

uvicorn==0.35.0

Werkzeug==3.1.3

python-dotenv

databases[postgresql]

sqlalchemy

asyncpg

psycopg2

pydantic[email]
websockets==12.0
wsproto==1.2.0

aio-pika==9.4.0
